// packages/measurements/src/measure.types.ts
interface PointMeasurement {
  position: Cesium.Cartesian3
  height: number
  slope: number
  drawStatus: number
  show: boolean
}